Adapting PDKHttpTransmitter to be more easily subclassed.
* Cleaning up Clang warnings.

+++ PassiveDataKit/PDKHttpTransmitter.m
@@ -9,18 +9,9 @@
 #import <SystemConfiguration/SystemConfiguration.h>
 #import <sqlite3.h>
 
-#import "NSString+RAInflections.h"
-
 #import "PDKAFURLSessionManager.h"
 
 #import "PDKHttpTransmitter.h"
-
-#define PDK_METADATA_KEY @"passive-data-metadata"
-#define PDK_GENERATOR_ID_KEY @"generator-id"
-#define PDK_GENERATOR_KEY @"generator"
-#define PDK_TIMESTAMP_KEY @"timestamp"
-
-#define PDK_DEFAULT_TRANSMITTER_ID @"http-transmitter"
 
 typedef enum {
     ConnectionTypeUnknown,
@@ -31,9 +22,6 @@ typedef enum {
 
 @interface PDKHttpTransmitter ()
 
-@property NSString * source;
-@property NSString * transmitterId;
-@property NSURL * uploadUrl;
 @property BOOL requireCharging;
 @property BOOL requireWiFi;
 
@@ -177,6 +165,21 @@ typedef enum {
     [self transmitReadings:0 completionHandler:completionHandler];
 }
 
+- (NSURLRequest *) uploadRequestForPayload:(NSArray *) payload {
+    NSMutableURLRequest * request = [[PDKAFJSONRequestSerializer serializer] requestWithMethod:@"CREATE"
+                                                                                     URLString:[self.uploadUrl description]
+                                                                                    parameters:payload
+                                                                                         error:nil];
+    [request setValue:@"application/json" forHTTPHeaderField:@"Content-Type"];
+    [request setValue:@"application/json" forHTTPHeaderField:@"Accept"];
+    
+    return request;
+}
+
+- (NSUInteger) payloadSize {
+    return 16;
+}
+
 - (void) transmitReadings:(NSUInteger) uploadWindow completionHandler:(void (^)(UIBackgroundFetchResult result)) completionHandler {
     if (uploadWindow == 0) {
         uploadWindow = 8; //!OCLINT
@@ -186,7 +189,7 @@ typedef enum {
     
     sqlite3_stmt * statement = NULL;
     
-    NSString * querySQL = @"SELECT D.id, D.properties FROM data D WHERE (D.timestamp < ?) LIMIT 16";
+    NSString * querySQL = [NSString stringWithFormat:@"SELECT D.id, D.properties FROM data D WHERE (D.timestamp < ?) LIMIT %d", (int) [self payloadSize]];
     
     const char * query_stmt = [querySQL UTF8String];
     
@@ -224,65 +227,62 @@ typedef enum {
         
         sqlite3_finalize(statement);
         
-        NSMutableURLRequest *req = [[PDKAFJSONRequestSerializer serializer] requestWithMethod:@"CREATE"
-                                                                                    URLString:[self.uploadUrl description]
-                                                                                   parameters:payload
-                                                                                        error:nil];
-        [req setValue:@"application/json" forHTTPHeaderField:@"Content-Type"];
-        [req setValue:@"application/json" forHTTPHeaderField:@"Accept"];
+        NSURLRequest * request = [self uploadRequestForPayload:payload];
         
-        NSLog(@"UPLOADING PAYLOAD: %@", payload);
-        
-        PDKAFURLSessionManager * manager = [[PDKAFURLSessionManager alloc] initWithSessionConfiguration:[NSURLSessionConfiguration defaultSessionConfiguration]];
-        
-        [[manager dataTaskWithRequest:req
-                       uploadProgress:nil
-                     downloadProgress:nil
-                    completionHandler:^(NSURLResponse * _Nonnull response, id  _Nullable responseObject, NSError * _Nullable error) {
-                        if ([responseObject containsObject:@"Data bundle added successfully, and ready for processing."] == NO) {
-                            NSLog(@"Invalid response: %@", responseObject);
-                        }
-                        else if (error == nil) {
-                            for (NSNumber * identifier in uploaded) {
-                                sqlite3_stmt * deleteStatement = NULL;
-                                
-                                NSString * deleteSQL = @"DELETE FROM data WHERE (id = ?)";
-                                
-                                const char * delete_stmt = [deleteSQL UTF8String];
-                                
-                                if (sqlite3_prepare_v2(self.database, delete_stmt, -1, &deleteStatement, NULL) == SQLITE_OK)
-                                {
-                                    sqlite3_bind_int(deleteStatement, 1, [identifier intValue]);
+        if (request != nil) {
+            NSLog(@"UPLOADING PAYLOAD: %@", payload);
+            
+            PDKAFURLSessionManager * manager = [[PDKAFURLSessionManager alloc] initWithSessionConfiguration:[NSURLSessionConfiguration defaultSessionConfiguration]];
+            
+            [[manager dataTaskWithRequest:request
+                           uploadProgress:nil
+                         downloadProgress:nil
+                        completionHandler:^(NSURLResponse * _Nonnull response, id  _Nullable responseObject, NSError * _Nullable error) {
+                            if ([responseObject containsObject:@"Data bundle added successfully, and ready for processing."] == NO) {
+                                NSLog(@"Invalid response: %@", responseObject);
+                            }
+                            else if (error == nil) {
+                                for (NSNumber * identifier in uploaded) {
+                                    sqlite3_stmt * deleteStatement = NULL;
                                     
-                                    while (sqlite3_step(deleteStatement) == SQLITE_ROW) { //!OCLINT
+                                    NSString * deleteSQL = @"DELETE FROM data WHERE (id = ?)";
+                                    
+                                    const char * delete_stmt = [deleteSQL UTF8String];
+                                    
+                                    if (sqlite3_prepare_v2(self.database, delete_stmt, -1, &deleteStatement, NULL) == SQLITE_OK)
+                                    {
+                                        sqlite3_bind_int(deleteStatement, 1, [identifier intValue]);
                                         
+                                        while (sqlite3_step(deleteStatement) == SQLITE_ROW) { //!OCLINT
+                                            
+                                        }
+                                        
+                                        sqlite3_finalize(deleteStatement);
+                                    } else {
+                                        if (completionHandler != nil) {
+                                            completionHandler(UIBackgroundFetchResultFailed);
+                                        }
+                                        
+                                        NSLog(@"Error while deleting data. '%s'", sqlite3_errmsg(self.database));
+                                        
+                                        return;
                                     }
-                                    
-                                    sqlite3_finalize(deleteStatement);
+                                }
+                                
+                                NSTimeInterval interval = [NSDate date].timeIntervalSince1970 - now;
+                                
+                                if (uploadWindow > 0 && interval > uploadWindow && uploaded.count > 0) {
+                                    [self transmitReadings:(uploadWindow - interval) completionHandler:completionHandler];
                                 } else {
                                     if (completionHandler != nil) {
-                                        completionHandler(UIBackgroundFetchResultFailed);
+                                        completionHandler(UIBackgroundFetchResultNewData);
                                     }
-
-                                    NSLog(@"Error while deleting data. '%s'", sqlite3_errmsg(self.database));
-
-                                    return;
+                                    
                                 }
                             }
                             
-                            NSTimeInterval interval = [NSDate date].timeIntervalSince1970 - now;
-                            
-                            if (uploadWindow > 0 && interval > uploadWindow && uploaded.count > 0) {
-                                [self transmitReadings:(uploadWindow - interval) completionHandler:completionHandler];
-                            } else {
-                                if (completionHandler != nil) {
-                                    completionHandler(UIBackgroundFetchResultNewData);
-                                }
-
-                            }
-                        }
-                        
-                    }] resume];
+                        }] resume];
+        }
     }
 }
 
@@ -294,9 +294,9 @@ typedef enum {
     return -1;
 }
 
-- (void) receivedData:(NSDictionary *) data forGenerator:(PDKDataGenerator) dataGenerator {
-    NSMutableDictionary * toStore = [NSMutableDictionary dictionaryWithDictionary:data];
-    
+- (NSDictionary *) processIncomingDataPoint:(NSDictionary *) dataPoint forGenerator:(PDKDataGenerator) dataGenerator {
+    NSMutableDictionary * toStore = [NSMutableDictionary dictionaryWithDictionary:dataPoint];
+
     if (toStore[PDK_METADATA_KEY] == nil) {
         NSMutableDictionary * metadata = [NSMutableDictionary dictionary];
         
@@ -310,30 +310,41 @@ typedef enum {
         } else {
             metadata[PDK_SOURCE_KEY] = [[PassiveDataKit sharedInstance] identifierForUser];
         }
-
+        
         metadata[PDK_TIMESTAMP_KEY] = [NSNumber numberWithDouble:[NSDate date].timeIntervalSince1970];
         
         toStore[PDK_METADATA_KEY] = metadata;
     }
+
+    return toStore;
+}
+
+
+- (void) receivedData:(NSDictionary *) dataPoint forGenerator:(PDKDataGenerator) dataGenerator {
+    NSDictionary * toStore = [self processIncomingDataPoint:dataPoint forGenerator:dataGenerator];
     
-    sqlite3_stmt * stmt;
-    
-    NSString * insert = @"INSERT INTO data (timestamp, properties) VALUES (?, ?);";
-    
-    if(sqlite3_prepare_v2(self.database, [insert UTF8String], -1, &stmt, NULL) == SQLITE_OK) {
-        sqlite3_bind_double(stmt, 1, [toStore[PDK_METADATA_KEY][PDK_TIMESTAMP_KEY] doubleValue]);
+    if (toStore != nil) {
+        NSLog(@"STORING: %@", toStore);
         
-        NSError * err = nil;
-        NSData * jsonData = [NSJSONSerialization dataWithJSONObject:toStore options:0 error:&err];
-        NSString * jsonString = [[NSString alloc] initWithData:jsonData encoding:NSUTF8StringEncoding];
+        sqlite3_stmt * stmt;
         
-        sqlite3_bind_text(stmt, 2, [jsonString UTF8String], -1, SQLITE_TRANSIENT);
+        NSString * insert = @"INSERT INTO data (timestamp, properties) VALUES (?, ?);";
         
-        if (SQLITE_DONE != sqlite3_step(stmt)) {
-            NSLog(@"Error while inserting data. '%s'", sqlite3_errmsg(self.database));
+        if(sqlite3_prepare_v2(self.database, [insert UTF8String], -1, &stmt, NULL) == SQLITE_OK) {
+            sqlite3_bind_double(stmt, 1, [toStore[PDK_METADATA_KEY][PDK_TIMESTAMP_KEY] doubleValue]);
+            
+            NSError * err = nil;
+            NSData * jsonData = [NSJSONSerialization dataWithJSONObject:toStore options:0 error:&err];
+            NSString * jsonString = [[NSString alloc] initWithData:jsonData encoding:NSUTF8StringEncoding];
+            
+            sqlite3_bind_text(stmt, 2, [jsonString UTF8String], -1, SQLITE_TRANSIENT);
+            
+            if (SQLITE_DONE != sqlite3_step(stmt)) {
+                NSLog(@"Error while inserting data. '%s'", sqlite3_errmsg(self.database));
+            }
+            
+            sqlite3_finalize(stmt);
         }
-        
-        sqlite3_finalize(stmt);
     }
 }
